requirements.txt
openai
semantic_kernel
instructor
python-dotenv